requirements.txt
# Audio capture
pyaudiowpatch>=0.2.12
sounddevice>=0.4.6
numpy>=1.24

# VAD
silero-vad>=5.1
torch>=2.0
torchaudio>=2.0

# Transcription — Whisper
faster-whisper>=1.0.0

# Transcription — GigaAM (optional)
# pip install git+https://github.com/salute-developers/GigaAM.git

# Diarization
pyannote.audio>=3.1

# Config
pyyaml>=6.0

# CLI
click>=8.0
